refactor : 테스트 클래스 분리
